requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via
    #   httpx
    #   openai
    #   starlette
    #   watchfiles
cachetools==5.3.3
    # via google-auth
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   typer
    #   uvicorn
colorama==0.4.6
    # via
    #   click
    #   pytest
    #   tqdm
    #   uvicorn
distro==1.9.0
    # via openai
dnspython==2.6.1
    # via email-validator
email-validator==2.1.1
    # via fastapi
fastapi==0.111.0
    # via
    #   -r requirements.in
    #   fastapi-cli
fastapi-cli==0.0.3
    # via fastapi
google-api-core==2.18.0
    # via google-api-python-client
google-api-python-client==2.126.0
    # via -r requirements.in
google-auth==2.29.0
    # via
    #   -r requirements.in
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
google-auth-httplib2==0.2.0
    # via
    #   -r requirements.in
    #   google-api-python-client
google-auth-oauthlib==1.2.0
    # via -r requirements.in
googleapis-common-protos==1.63.0
    # via google-api-core
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.5
    # via httpx
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httptools==0.6.1
    # via uvicorn
httpx==0.27.0
    # via
    #   fastapi
    #   openai
idna==3.7
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
iniconfig==2.0.0
    # via pytest
jinja2==3.1.4
    # via fastapi
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
mypy==1.10.0
    # via -r requirements.in
mypy-extensions==1.0.0
    # via mypy
oauthlib==3.2.2
    # via requests-oauthlib
openai==1.30.1
    # via -r requirements.in
orjson==3.10.3
    # via fastapi
packaging==24.0
    # via pytest
pluggy==1.5.0
    # via pytest
proto-plus==1.23.0
    # via google-api-core
protobuf==4.25.3
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
pyasn1==0.6.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.0
    # via google-auth
pydantic==2.7.0
    # via
    #   fastapi
    #   openai
pydantic-core==2.18.1
    # via pydantic
pygments==2.18.0
    # via rich
pyparsing==3.1.2
    # via httplib2
pytest==8.2.1
    # via -r requirements.in
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   uvicorn
python-multipart==0.0.9
    # via fastapi
pyyaml==6.0.1
    # via uvicorn
requests==2.31.0
    # via
    #   google-api-core
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rich==13.7.1
    # via typer
rsa==4.9
    # via google-auth
ruff==0.4.4
    # via -r requirements.in
shellingham==1.5.4
    # via typer
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
starlette==0.37.2
    # via fastapi
tqdm==4.66.4
    # via openai
typer==0.12.3
    # via fastapi-cli
typing-extensions==4.11.0
    # via
    #   fastapi
    #   mypy
    #   openai
    #   pydantic
    #   pydantic-core
    #   typer
ujson==5.10.0
    # via fastapi
uritemplate==4.1.1
    # via google-api-python-client
urllib3==2.2.1
    # via requests
uvicorn[standard]==0.24.0
    # via
    #   -r requirements.in
    #   fastapi
    #   fastapi-cli
watchfiles==0.21.0
    # via uvicorn
websockets==12.0
    # via uvicorn
